15 - deleted viewController and added all methods in bikeController
